rearange imports so can use WorkflowConfigParser without pegasus installed
